Move and disable auto-minor-mode
It was causing problems with bootstrapping.  Once everything is nice and
ready as far as installed packages, the use-package form can be
re-enabled (if desired).  The reason this fails is because
`package-refresh-contents' uses `find-file' (presumably) to do its
thing.  Since the function was on a hook, I couldn't find any files
until this function was defined!

It's likely the *real* fix in this commit is the custom-file change.  I
may re-enable the hook at a later time.